Some refactoring for clarity per Nicks' code review
